refactor(extract): more symmetry
